tools/scd-validator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scs-validator"
version = "0.1.0"
description = "Validator for Structured Context Specification (SCS) documents and bundles"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Timothy McCrimmon", email = "tim@ohana-consulting.com"}
]
keywords = ["scs", "validation", "structured-context", "ai-native"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.1",
    "colorama>=0.4",
    "tabulate>=0.9",
    "semver>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "mypy>=1.7",
    "ruff>=0.1",
]

[project.scripts]
scs = "scs_validator.cli:cli"

[project.urls]
Homepage = "https://github.com/tim-mccrimmon/scs-spec"
Repository = "https://github.com/tim-mccrimmon/scs-spec"
Issues = "https://github.com/tim-mccrimmon/scs-spec/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/scs_validator --cov-report=term-missing"
